refactoring navigation
