+++ navigation/StackNavigator.js
@@ -1,0 +1,33 @@
+
+  
+import React from "react";
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
+import { BottomTabNavigator } from './TabNavigator'
+
+const Stack = createNativeStackNavigator();
+
+const screenOptionStyle = {
+  headerStyle: {
+    backgroundColor: "#9AC4F8",
+  },
+  headerTintColor: "white",
+  headerBackTitle: "Back",
+};
+
+const MainStackNavigator = () => {
+  return (
+    <Stack.Navigator
+        initialRouteName="Settings"
+        screenOptions={{
+          headerStyle: { backgroundColor: '#633689' },
+          headerTintColor: '#fff',
+          headerTitleStyle: { fontWeight: 'bold' },
+          headerShown: false,
+        }}>
+        <Stack.Screen name="BottomTabNavigator" component={BottomTabNavigator} options={{ title: 'Tab Navigation' }}/>
+      </Stack.Navigator>
+  );
+};
+
+
+export { MainStackNavigator };
